Time: 8 ms (68.21%), Space: 48 MB (30.1%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -14,7 +14,7 @@ class Solution {
             long hours = 0;
 
             for (int pile : piles) {
-                hours += (pile + mid - 1L) / mid;
+                hours += (long)(pile + mid - 1) / mid;
             }
 
             if (hours <= h) {
